Python_OCR_JE/requirements.txt
pillow
pytesseract
opencv-python